fix: Resolve temporal workflows module resolution errors
- Create shared encryption utilities to consolidate password handling
- Fix TypeScript path aliases in temporal workflows using custom build script
- Update tsconfig to properly resolve @shared/* imports
- Replace tsc-alias with custom fix-imports.js script for reliable path resolution
- Migrate all password generation/hashing to use shared utilities
- Update auth verification endpoint to use new secret provider methods

This resolves the ERR_MODULE_NOT_FOUND errors when importing @shared/db
and other shared modules in the temporal workflows package.
